Removed @Inject, redundant empty lines, renamed the class, used full name of controllers as urls, changed the type of product's price(from Long to BigDecimal)

+++ src/main/java/mate/academy/internetshop/model/Product.java
@@ -1,11 +1,13 @@
 package mate.academy.internetshop.model;
 
+import java.math.BigDecimal;
+
 public class Product {
     private Long id;
     private String name;
-    private Double price;
+    private BigDecimal price;
 
-    public Product(String name, double price) {
+    public Product(String name, BigDecimal price) {
         this.name = name;
         this.price = price;
     }
@@ -26,11 +28,11 @@ public class Product {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
